Replace all Centered_Text instances with Text
Newly made Text class is to now be used for all instances of text.
